Отрисовка блока путешествия добавлена в метод рендеринга
